improvement(Payroll) - Break multipayroll/index.js in multiple files - Improve the calcul of workedDays in use external function to get range   of dates between two Dates
