+++ client/src/modules/employees/registration/employees.html
@@ -303,22 +303,6 @@
                       </div>
                     </div>
                   </div>
-                  <!-- Employee number of spouse-->
-                  <!-- <div class="form-group"
-                    ng-class="{'has-error' : OptionalForm.nb_spouse.$invalid && EmployeeRegistrationForm.$submitted}">
-                    <label class="col-md-3 control-label" translate>FORM.LABELS.NB_SPOUSE</label>
-                    <div class="col-md-9">
-                      <input
-                        type="number"
-                        class="form-control"
-                        name="nb_spouse"
-                        ng-model="EmployeeCtrl.employee.nb_spouse"
-                        autocomplete="off">
-                     <div class="help-block" ng-messages="OptionalForm.nb_spouse.$error" ng-show="EmployeeRegistrationForm.$submitted">
-                      <div ng-messages-include="modules/templates/messages.tmpl.html"></div>
-                     </div>
-                    </div>
-                  </div> -->
 
                   <!-- Employee number of child-->
                   <div class="form-group"
